+++ SEPPmail365cloud.psm1
@@ -23,7 +23,8 @@ if ($sc365notest -ne $true) {
         Write-Warning "You are not connected to Exchange Online. Use `"Connect-ExchangeOnline`" to connect to your tenant"
     } else {
         try {
-            if ((Get-OrganizationConfig).IsDehydrated) {
+            $isDehydrated = (Get-OrganizationConfig).IsDehydrated
+            if ($isDehydrated) {
                 Write-Verbose "Organisation is not enabled for customizations -- is 'Dehyrated'. Turning this on now"
                 Enable-OrganizationCustomization  #-confirm:$false
             }